requirements.txt
numpy>=1.26.3
pocketsphinx>=5.0.3
PyAudio>=0.2.14
pynput>=1.7.6
PySide6>=6.6.1
PySide6_Addons>=6.6.1
PySide6_Essentials>=6.6.1
Requests>=2.31.0
shiboken6>=6.6.1
soundfile>=0.12.1
SpeechRecognition>=3.10.1
torch>=2.1.2
vosk>=0.3.45
openai-whisper>=1.1.10
